extracts bookmark logic; cleans up login status
- extract bookmark save logic to separate file
- ensure auth_token is cleared from local cache on logout
